Separate passphrase to another file and split word generation
