+++ crates/bitwarden/src/tool/generators/mod.rs
@@ -1,4 +1,6 @@
 mod client_generator;
+mod passphrase;
 mod password;
 
-pub use password::{PassphraseGeneratorRequest, PasswordGeneratorRequest};
+pub use passphrase::PassphraseGeneratorRequest;
+pub use password::PasswordGeneratorRequest;
